Moved pure math functions to their own lib
